Parse HTTP request?
Check to make sure the parameter r is not empty or null. Break up the request into string tokens using the String split() method. Use the argument “[ \t\n?]” to split the string if a space, tab, newline, or question mark are encountered. Verify there are at least two tokens. Verify the first token is ‘GET’. The second token should be the file path. Make sure it is not null or empty. Set the three instance variables.